+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_help.py
@@ -442,8 +442,8 @@ helps['acr import'] = """
             az acr import -n Myregistry --source MyRegistry.azurecr.io/repository:tag -t targetrepository:targettag
         - name: Import an image from a registry in the same subscription as the target registry is in.
           text: >
-            az acr import -n MyRegistry --source sourceregistry.azurecr.io/repository:tag -t targetrepository:targettag
+            az acr import -n MyRegistry --source repository:tag -r sourceregistry -t targetrepository:targettag
         - name: Import an image from a registry in a different subscription from the one the target registry is in.
           text: >
-            az acr import -n MyRegistry --source sourceregistry.azurecr.io/repository:tag -t targetrepository:targettag -r /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/sourceResourceGroup/providers/Microsoft.ContainerRegistry/registries/sourceRegistry
+            az acr import -n MyRegistry --source repository:tag -t targetrepository:targettag -r /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/sourceResourceGroup/providers/Microsoft.ContainerRegistry/registries/sourceRegistry
 """